+++ src/components/best-sum.c
@@ -1,4 +1,4 @@
-#include "urn-component.h"
+#include "last-component.h"
 
 typedef struct _UrnBestSum {
     UrnComponent base;
