+++ src/last.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <time.h>
 #include <jansson.h>
-#include "headers/urn.h"
+#include "headers/last.h"
 
 long long urn_time_now(void) {
     struct timespec timespec;
